Fix lifecycle handler, update Crux
Update Crux to the newest version with better Kafka support.
Fixed a bug which occurs when uninstalling addons. Atlassian will send 2
lifecycle calls:

1. Disabled
2. Uninstalled

But it does so without waiting for the previous one to return so we need
to retry when we save because we are almost always looking at a failing
CAS operation.

Resolves: ATOP-60